vertice2/requirements.txt
autopep8==1.7.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
dnspython==2.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycodestyle==2.9.1
pymongo==4.2.0
python-dotenv==0.21.0
requests==2.28.2
six==1.16.0
toml==0.10.2
urllib3==1.26.14
Werkzeug==2.2.2
